Fixed a dumb issue, met MVP?
A component is changing a controlled input of type password to be uncontrolled. Input elements should not switch from controlled to uncontrolled (or vice versa). Decide between using a controlled or uncontrolled input element for the lifetime of the component. More info: https://fb.me/react-controlled-components

basically i has setState instead of state.  shoutout to my girlfriend who was able to look at this and see the issue who has NO CODING EXPERIENCE.

good news. i can now log in and out of the server and see the protected page with this. so im at mvp i think?